Break sensor.py into multiple files
