+++ simplipy/sensor/v2.py
@@ -1,9 +1,8 @@
-"""Define a SimpliSafe sensor."""
+"""Define a v2 (old) SimpliSafe sensor."""
 import logging
-from typing import Optional
 
-from .entity import Entity, EntityTypes, EntityV3
-from .errors import SimplipyError
+from simplipy.entity import Entity, EntityTypes
+from simplipy.errors import SimplipyError
 
 _LOGGER: logging.Logger = logging.getLogger(__name__)
 
@@ -43,36 +42,3 @@ class SensorV2(Entity):
             return self.entity_data.get("entryStatus", "closed") == "open"
 
         raise SimplipyError(f"Cannot determine triggered state for sensor: {self.name}")
-
-
-class SensorV3(EntityV3):
-    """Define a V3 (new) sensor."""
-
-    @property
-    def trigger_instantly(self) -> bool:
-        """Return whether the sensor will trigger instantly."""
-        return self.entity_data["setting"]["instantTrigger"]
-
-    @property
-    def triggered(self) -> bool:
-        """Return the sensor's status info."""
-        if self.type in (
-            EntityTypes.carbon_monoxide,
-            EntityTypes.entry,
-            EntityTypes.glass_break,
-            EntityTypes.leak,
-            EntityTypes.motion,
-            EntityTypes.smoke,
-            EntityTypes.temperature,
-        ):
-            return self.entity_data["status"].get("triggered", False)
-
-        return False
-
-    @property
-    def temperature(self) -> Optional[int]:
-        """Return the sensor's status info."""
-        if self.type != EntityTypes.temperature:
-            raise AttributeError("Non-temperature sensor cannot have a temperature")
-
-        return self.entity_data["status"]["temperature"]
